Don't remove temporary files if debug enabled

+++ cleanupContext.sh
@@ -2,7 +2,7 @@
 
 if [[ -n "${GSGEN_DEBUG}" ]]; then set ${GSGEN_DEBUG}; fi
 
-if [[ -n "${ROOT_DIR}" ]]; then
+if [[ (-z "${GSGEN_DEBUG}") && (-n "${ROOT_DIR}") ]]; then
     find ${ROOT_DIR} -name "aggregate_*.json" -delete
     find ${ROOT_DIR} -name "STATUS.txt" -delete
     find ${ROOT_DIR} -name "stripped_*.json" -delete
